Make ProgressBar usable in more places than just MusicController
